[fixed] stability fixes from TC [added] cell movement on all planets
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3devel/MMOCoreORB@371 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/objects/auction/AuctionController.h
@@ -92,7 +92,7 @@ public:
 		return item;
 	}
 	
-	void getBazaarData(Player* player, long long objectid, int screen, int extent, unsigned int category, int count, int offset) {
+	void getBazaarData(Player* player, uint64 objectid, int screen, int extent, unsigned int category, int count, int offset) {
 		lock();
 		
 		AuctionQueryHeadersResponseMessage* reply = new AuctionQueryHeadersResponseMessage(screen, count);
